// src/components/ui/Input/input.enums.ts
export enum InputVariant{
    FILLED = "filled",
    OUTLINED = "outlined",
    STANDARD = "standard"
}

export enum InputType {
    TEXT = "text",
    PASSWORD = "password",
    EMAIL = "email",
    NUMBER = "number",
    TEL = "tel",
    URL = "url",
    SEARCH = "search",
    DATE = "date",
    DATETIME_LOCAL = "datetime-local",
    MONTH = "month",
    WEEK = "week",
    TIME = "time",
    COLOR = "color",
    FILE = "file",
    CHECKBOX = "checkbox",
    RADIO = "radio",
    RANGE = "range",
    HIDDEN = "hidden",
    SUBMIT = "submit",
    RESET = "reset",
    BUTTON = "button",
  }
  